requirements.txt
azure-ai-contentsafety==1.0.0
aiohttp>=3.9.0
azure-identity==1.19.0
azure-mgmt-cognitiveservices==13.5.0
azure-mgmt-resource==23.1.0
azure-ai-projects==1.0.0b10
openai==1.55.0
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
python-multipart==0.0.12
Pillow==11.0.0
aiofiles==24.1.0
python-pptx>=1.0.0
